+++ nodes/ValidatorNode/types.ts
@@ -221,3 +221,35 @@ export interface InputField {
     vatCountryCodeCustom?: string;
     taxIdLocaleCustom?: string;
 }
+
+export interface PhoneRewriteInput {
+    source: string;
+    expectedTypes?: Array<
+        | 'FIXED_LINE'
+        | 'MOBILE'
+        | 'FIXED_LINE_OR_MOBILE'
+        | 'TOLL_FREE'
+        | 'PREMIUM_RATE'
+        | 'SHARED_COST'
+        | 'VOIP'
+        | 'PERSONAL_NUMBER'
+        | 'PAGER'
+        | 'UAN'
+        | 'VOICEMAIL'
+        | 'UNKNOWN'
+    >;
+    fallbackTypes?: Array<
+        | 'FIXED_LINE'
+        | 'MOBILE'
+        | 'FIXED_LINE_OR_MOBILE'
+        | 'TOLL_FREE'
+        | 'PREMIUM_RATE'
+        | 'SHARED_COST'
+        | 'VOIP'
+        | 'PERSONAL_NUMBER'
+        | 'PAGER'
+        | 'UAN'
+        | 'VOICEMAIL'
+        | 'UNKNOWN'
+    >;
+}